requirements.txt
fastapi==0.75.2
requests==2.26.0
uvicorn==0.17.6
dnspython==2.1.0
beanie==1.11.7
motor==3.0.0